bugfixed

+++ Amd.sh
@@ -10,20 +10,29 @@ miner=7
 # Update system path
 apt-get update && apt-get upgrade
 # Install necessary package
-apt-get install sudo vim wget curl systemd git opencl-headers ocl-icd-libopencl1 ocl-icd-opencl-dev
+apt-get install sudo vim wget curl systemd git ntp opencl-headers ocl-icd-libopencl1 ocl-icd-opencl-dev
 # Set mining folder
-sudo mkdir -p /opt/ton-miner/
+if [[ ! -f "/opt/ton-miner/" ]]; then
+    sudo mkdir -p /opt/ton-miner/
+fi
 cd /opt/ton-miner/
 # Download mining package
-wget https://github.com/tontechio/pow-miner-gpu/releases/download/20211112-3/minertools-opencl-ubuntu-18.04-x86-64.tar.gz
-# Unzip mining package
-tar xzf /opt/ton-miner/minertools-opencl-ubuntu-18.04-x86-64.tar.gz -C /opt/ton-miner/
-# Download Config (Need to re-download once a week)
-cd /opt/ton-miner && curl -L -O https://newton-blockchain.github.io/global.config.json
-chmod 777 global.config.json
+if [[ ! -e "minertools-opencl-ubuntu-18.04-x86-64.tar.gz" ]]; then
+    wget https://github.com/tontechio/pow-miner-gpu/releases/download/20211112-3/minertools-opencl-ubuntu-18.04-x86-64.tar.gz
+    # Unzip mining package
+    tar xzf /opt/ton-miner/minertools-opencl-ubuntu-18.04-x86-64.tar.gz -C /opt/ton-miner/
+fi
+if [[ ! -e "global.config.json" ]]; then
+    # Download Config (Need to re-download once a week)
+    cd /opt/ton-miner && curl -L -O https://newton-blockchain.github.io/global.config.json
+    chmod 777 global.config.json
+fi
 # Show mining log when start
 sed -i -e '$i \tail -f /var/log/syslog \n' /etc/rc.local
 # Crontab download global.config.json
+if [[ -e "/opt/ton-miner/DailyDownload.sh" ]]; then
+    rm -rf DailyDownload.sh
+fi
 wget https://raw.githubusercontent.com/odycenter/an-ton/main/DailyDownload.sh
 chmod 777 DailyDownload.sh
 echo "# Crontab download global.config.json" >> /etc/crontab
@@ -31,8 +40,9 @@ echo "* 15    * * *    root    bash /opt/ton-miner/DailyDownload.sh" >> /etc/cro
 # Write running parameter
 for i in $(seq 0 $miner);
 do
-# echo $ton_miner" -v 3 -C "$global_config" -e 'pminer start "$giver_address" "$my_address" "$i" 32 0'  [-l logfile]"
-# echo $ton_miner" -v 3 -C "$global_config" -e 'pminer start "$giver_address" "$my_address" "$i" 32 0'  [-l logfile]" >> /etc/rc.local
+if [[ -e "/etc/systemd/system/miner_gpu$i.service" ]]; then
+    rm -rf /etc/systemd/system/miner_gpu$i.service
+fi
 echo "[Unit]
 Description=TON miner
 After=network.target
