+++ Nvidia.sh
@@ -13,21 +13,33 @@ apt-get update && apt-get upgrade
 sudo mkdir -p /opt/ton-miner
 cd /opt/ton-miner
 # Install Nvidia cuda
-wget https://developer.download.nvidia.com/compute/cuda/repos/ubuntu1804/x86_64/cuda-ubuntu1804.pin
-sudo mv cuda-ubuntu1804.pin /etc/apt/preferences.d/cuda-repository-pin-600
-sudo apt-key adv --fetch-keys https://developer.download.nvidia.com/compute/cuda/repos/ubuntu1804/x86_64/7fa2af80.pub
-sudo add-apt-repository "deb https://developer.download.nvidia.com/compute/cuda/repos/ubuntu1804/x86_64/ /"
-apt-get install sudo vim wget curl systemd git nvidia-settings cuda-drivers-495 cuda-runtime-11-5 cuda-demo-suite-11-5 cuda-11-5 cuda
+echo "888"
+if [[ ! -e "/opt/ton-miner/cuda-ubuntu1804.pin" ]]; then
+echo "999"
+    apt-get install gnupg gnupg2 software-properties-common
+    wget https://developer.download.nvidia.com/compute/cuda/repos/ubuntu1804/x86_64/cuda-ubuntu1804.pin
+    sudo mv cuda-ubuntu1804.pin /etc/apt/preferences.d/cuda-repository-pin-600
+    sudo apt-key adv --fetch-keys https://developer.download.nvidia.com/compute/cuda/repos/ubuntu1804/x86_64/7fa2af80.pub
+    sudo add-apt-repository "deb https://developer.download.nvidia.com/compute/cuda/repos/ubuntu1804/x86_64/ /"
+    apt-get install sudo vim wget curl systemd git ntp  nvidia-settings cuda-drivers-495 cuda-runtime-11-5 cuda-demo-suite-11-5 cuda-11-5 cuda
+fi
 # Download mining package
-wget https://github.com/tontechio/pow-miner-gpu/releases/download/20211112-3/minertools-cuda-ubuntu-18.04-x86-64.tar.gz
-# Unzip mining package
-tar xzf /opt/ton-miner/minertools-cuda-ubuntu-18.04-x86-64.tar.gz -C /opt/ton-miner/
+if [[ ! -e "/opt/ton-miner/minertools-cuda-ubuntu-18.04-x86-64.tar.gz" ]]; then
+    wget https://github.com/tontechio/pow-miner-gpu/releases/download/20211112-3/minertools-cuda-ubuntu-18.04-x86-64.tar.gz
+    # Unzip mining package
+    tar xzf /opt/ton-miner/minertools-cuda-ubuntu-18.04-x86-64.tar.gz -C /opt/ton-miner/
+fi
 # Download Config (Need to re-download once a week)
-curl -L -O https://newton-blockchain.github.io/global.config.json
-chmod 777 global.config.json
+if [[ ! -e "/opt/ton-miner/global.config.json" ]]; then
+    curl -L -O https://newton-blockchain.github.io/global.config.json
+    chmod 777 global.config.json
+fi
 # Show mining log when start
 sed -i -e '$i \tail -f /var/log/syslog \n' /etc/rc.local
 # Crontab download global.config.json
+if [[ -e "/opt/ton-miner/DailyDownload.sh" ]]; then
+    rm -rf DailyDownload.sh
+fi
 wget https://raw.githubusercontent.com/odycenter/an-ton/main/DailyDownload.sh
 chmod 777 DailyDownload.sh
 echo "# Crontab download global.config.json" >> /etc/crontab
@@ -35,6 +47,9 @@ echo "* 15    * * *    root    bash /opt/ton-miner/DailyDownload.sh" >> /etc/cro
 # Write running parameter
 for i in $(seq 0 $miner);
 do
+if [[ -e "/etc/systemd/system/miner_gpu$i.service" ]]; then
+    rm -rf /etc/systemd/system/miner_gpu$i.service
+fi
 echo "[Unit]
 Description=TON miner
 After=network.target
